--- bin/launch_asynchronous.js
@@ -1,7 +1,0 @@
-require("../client");
-
-const AsynchronousWorkflow = require("../src/Workflows/AsynchronousWorkflow");
-
-new AsynchronousWorkflow().dispatch().catch(err => {
-  console.error(err);
-});
