--- bin/launch_error.js
@@ -1,7 +1,0 @@
-require("../client");
-
-const ErrorWorkflow = require("../src/Workflows/ErrorWorkflow");
-
-new ErrorWorkflow().dispatch().catch(err => {
-  console.error(err);
-});
