+++ bin/launch_node.js
@@ -1,0 +1,7 @@
+require("../client");
+
+const NodeWorkflow = require("../src/Workflows/NodeWorkflow");
+
+new NodeWorkflow().dispatch().catch(err => {
+  console.error(err);
+});
