--- bin/launch_parallel.js
@@ -1,7 +1,0 @@
-require("../client");
-
-const ParallelWorkflow = require("../src/Workflows/ParallelWorkflow");
-
-new ParallelWorkflow().dispatch().catch(err => {
-  console.error(err);
-});
