+++ bin/launch_python.js
@@ -1,0 +1,7 @@
+require("../client");
+
+const PythonWorkflow = require("../src/Workflows/PythonWorkflow");
+
+new PythonWorkflow().dispatch().catch(err => {
+  console.error(err);
+});
