--- bin/launch_sequential.js
@@ -1,7 +1,0 @@
-require("../client");
-
-const SequentialWorkflow = require("../src/Workflows/SequentialWorkflow");
-
-new SequentialWorkflow().dispatch().catch(err => {
-  console.error(err);
-});
