--- bin/launch_single_task.js
@@ -1,7 +1,0 @@
-require("../client");
-
-const TaskA = require("../src/Tasks/TaskA");
-
-new TaskA().dispatch().catch(err => {
-  console.error(err);
-});
