--- bin/launch_version.js
@@ -1,7 +1,0 @@
-require("../client");
-
-const VersionWorkflow = require("../src/Workflows/VersionWorkflow");
-
-new VersionWorkflow().dispatch().catch(err => {
-  console.error(err);
-});
