--- bin/launch_wait.js
@@ -1,7 +1,0 @@
-require("../client");
-
-const WaitWorkflow = require("../src/Workflows/WaitWorkflow");
-
-new WaitWorkflow().dispatch().catch(err => {
-  console.error(err);
-});
