--- bin/launch_wait_event.js
@@ -1,17 +1,0 @@
-require("../client");
-
-const uniqid = require("uniqid");
-const WaitEventWorkflow = require("../src/Workflows/WaitEventWorkflow");
-
-const id = uniqid();
-new WaitEventWorkflow(id).dispatch().catch(err => {
-  console.error(err);
-});
-
-setTimeout(function() {
-  WaitEventWorkflow.whereId(id)
-    .send("MyEvent", {})
-    .catch(err => {
-      console.error(err);
-    });
-}, 1000);
