+++ src/Tests/Node/DemoWorkflow.js
@@ -1,8 +1,8 @@
 const { Workflow } = require("zenaton");
-const TaskA = require("../Tasks/TaskA");
-const TaskB = require("../Tasks/TaskB");
-const TaskC = require("../Tasks/TaskC");
-const TaskD = require("../Tasks/TaskD");
+const TaskA = require("../../Tasks/TaskA");
+const TaskB = require("../../Tasks/TaskB");
+const TaskC = require("../../Tasks/TaskC");
+const TaskD = require("../../Tasks/TaskD");
 
 module.exports = Workflow("SequentialWorkflow", async function() {
   const a = await new TaskA().execute();
