--- test/output/launch_asynchronous.out
@@ -1,8 +1,0 @@
-Task A starts
-Task B starts
-Task C starts
-Task A ends
-Task B ends
-Task C ends
-Task D starts
-Task D ends
