--- test/output/launch_error.out
@@ -1,3 +1,0 @@
-Task A starts
-Task E starts
-Task A ends
