--- test/output/launch_event.out
@@ -1,4 +1,0 @@
-Task A starts
-Task A ends
-Task C starts
-Task C ends
