--- test/output/launch_parallel.out
@@ -1,6 +1,0 @@
-Task A starts
-Task B starts
-Task A ends
-Task B ends
-Task D starts
-Task D ends
