--- test/output/launch_recursive.out
@@ -1,18 +1,0 @@
-0
-1
-2
-3
-4
-5
-6
-7
-8
-9
-
-Iteration: 1
-0
-1
-2
-3
-4
-5
