--- test/output/launch_sequential.out
@@ -1,25 +1,0 @@
-6
-Task A starts
-7
-Task A ends
-8
-Task C starts
-9
-
-Iteration: 2
-0
-1
-Task C ends
-Task D starts
-2
-3
-4
-5
-6
-Task D ends
-7
-8
-9
-
-Iteration: 3
-0
