--- test/output/launch_single_task.out
@@ -1,20 +1,0 @@
-Task A starts
-1
-2
-Task A ends
-3
-4
-5
-6
-7
-8
-9
-
-Iteration: 4
-0
-1
-2
-3
-4
-5
-6
