--- test/output/launch_version.out
@@ -1,27 +1,0 @@
-Task A starts
-Task B starts
-Task C starts
-Task D starts
-7
-8
-Task A ends
-9
-Task B ends
-
-Iteration: 5
-Task C ends
-0
-Task D ends
-1
-2
-3
-4
-5
-6
-7
-8
-9
-
-Iteration: 6
-0
-1
