--- test/output/launch_wait.out
@@ -1,21 +1,0 @@
-Task A starts
-2
-3
-Task A ends
-4
-5
-6
-Task B starts
-7
-8
-Task B ends
-9
-
-Iteration: 7
-0
-1
-2
-3
-4
-5
-6
