--- test/output/launch_wait_event.out
@@ -1,21 +1,0 @@
-8
-Task A starts
-9
-Task A ends
-
-Iteration: 8
-0
-1
-2
-3
-4
-5
-6
-7
-8
-9
-
-Iteration: 9
-0
-1
-2
